Moved tests to a separated directory `tests`
