+++ src/tests/InfoModal.test.tsx
@@ -1,5 +1,5 @@
 import { render } from "@testing-library/react";
-import InfoModal from "./InfoModal";
+import InfoModal from "../modals/InfoModal";
 
 test("renders without crashing", () => {
   const { baseElement } = render(
