+++ src/tests/TabHome.test.tsx
@@ -1,5 +1,5 @@
 import { render } from "@testing-library/react";
-import TabDetails from "./TabDetails";
+import TabDetails from "../pages/TabDetails";
 
 test("renders without crashing", () => {
   const { baseElement } = render(<TabDetails />);
